Move builtin files inside folder
