+++ pkg/builtins/bitwise_test.go
@@ -1,10 +1,10 @@
-package builtinrunner_test
+package builtins_test
 
 import (
 	"testing"
 
+	"github.com/lambdaclass/cairo-vm.go/pkg/builtins"
 	"github.com/lambdaclass/cairo-vm.go/pkg/lambdaworks"
-	builtinrunner "github.com/lambdaclass/cairo-vm.go/pkg/runners/builtin_runner"
 	"github.com/lambdaclass/cairo-vm.go/pkg/vm/cairo_run"
 	"github.com/lambdaclass/cairo-vm.go/pkg/vm/memory"
 )
@@ -24,7 +24,7 @@ func TestDeduceMemoryCellBitwiseForPresetMemoryValidAnd(t *testing.T) {
 	mem.Memory.Insert(rel2, m2)
 	mem.Memory.Insert(rel3, m3)
 
-	builtin := builtinrunner.NewBitwiseBuiltinRunner(true)
+	builtin := builtins.NewBitwiseBuiltinRunner(true)
 
 	address := memory.NewRelocatable(0, 7)
 
@@ -56,7 +56,7 @@ func TestDeduceMemoryCellBitwiseForPresetMemoryValidXor(t *testing.T) {
 	mem.Memory.Insert(rel2, m2)
 	mem.Memory.Insert(rel3, m3)
 
-	builtin := builtinrunner.NewBitwiseBuiltinRunner(true)
+	builtin := builtins.NewBitwiseBuiltinRunner(true)
 
 	address := memory.NewRelocatable(0, 8)
 
@@ -88,7 +88,7 @@ func TestDeduceMemoryCellBitwiseForPresetMemoryValidOr(t *testing.T) {
 	mem.Memory.Insert(rel2, m2)
 	mem.Memory.Insert(rel3, m3)
 
-	builtin := builtinrunner.NewBitwiseBuiltinRunner(true)
+	builtin := builtins.NewBitwiseBuiltinRunner(true)
 
 	address := memory.NewRelocatable(0, 9)
 
@@ -120,7 +120,7 @@ func TestDeduceMemoryCellBitwiseForPresetMemoryIncorrectOffset(t *testing.T) {
 	mem.Memory.Insert(rel2, m2)
 	mem.Memory.Insert(rel3, m3)
 
-	builtin := builtinrunner.NewBitwiseBuiltinRunner(true)
+	builtin := builtins.NewBitwiseBuiltinRunner(true)
 
 	address := memory.NewRelocatable(0, 5)
 
@@ -148,7 +148,7 @@ func TestDeduceMemoryCellBitwiseForPresetMemoryNoValuesToOperate(t *testing.T) {
 	mem.Memory.Insert(rel1, m1)
 	mem.Memory.Insert(rel2, m2)
 
-	builtin := builtinrunner.NewBitwiseBuiltinRunner(true)
+	builtin := builtins.NewBitwiseBuiltinRunner(true)
 
 	address := memory.NewRelocatable(0, 5)
 
@@ -165,7 +165,8 @@ func TestDeduceMemoryCellBitwiseForPresetMemoryNoValuesToOperate(t *testing.T) {
 }
 
 func TestIntegrationBitwise(t *testing.T) {
-	_, err := cairo_run.CairoRun("../../../cairo_programs/bitwise_builtin_test.json")
+	t.Helper()
+	_, err := cairo_run.CairoRun("../../cairo_programs/bitwise_builtin_test.json")
 	if err != nil {
 		t.Errorf("fail to create cairo runner with error: %v", err)
 	}
